# api/swagger/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.TaskUsage:
    properties:
      model_tier:
        type: string
      task_id:
        type: string
      title:
        type: string
      token_usage:
        type: integer
    type: object
...
  api.UsageResponse:
    properties:
      project_id:
        type: string
      task_breakdown:
        items:
          $ref: '#/definitions/api.TaskUsage'
        type: array
      total_tokens:
        type: integer
    type: object
  orchestrator.Task:
    properties:
      assigned_agent_id:
        type: string
      assigned_role:
        type: string
      automation_eligible:
        type: boolean
      context_summary:
        type: string
      created_at:
        type: string
      depends_on:
        items:
          type: string
        type: array
      description:
        type: string
      id:
        type: string
      model_tier:
        type: string
      project_id:
        type: string
      risk_level:
        type: string
      spec_id:
        type: string
      status:
        type: string
      title:
        type: string
      token_usage:
        type: integer
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Spec-driven AI development platform — orchestrates teams of Claude
    Code agents.
  title: Foundry API
  version: 0.1.0
paths:
...
  /projects/{id}/tasks:
    get:
      description: Returns all tasks for the given project; optionally filter by status
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by task status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/orchestrator.Task'
            type: array
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List tasks
      tags:
      - tasks
  /projects/{id}/tasks/{taskId}:
    get:
      description: Returns the task with the given ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/orchestrator.Task'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get task
      tags:
      - tasks
  /projects/{id}/usage:
    get:
      description: Returns token usage aggregated across all tasks for the given project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UsageResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get usage
      tags:
      - usage
produces:
- application/json
schemes:
- http
swagger: "2.0"